PS-2-1 сделал отступы, toFloat

+++ src/main/kotlin/lesson2/lesson2_test1.kt
@@ -1,13 +1,13 @@
 package org.example.lesson2
 
-fun main(){
+fun main() {
 
     val ratingOFStudent1 = 3
     val ratingOFStudent2 = 4
     val ratingOFStudent3 = 3
     val ratingOFStudent4 = 5
 
-    val ratingAverage = (ratingOFStudent1 + ratingOFStudent2 + ratingOFStudent3 + ratingOFStudent4)/4f
+    val ratingAverage = (ratingOFStudent1 + ratingOFStudent2 + ratingOFStudent3 + ratingOFStudent4).toFloat() / 4
 
     println(ratingAverage)
 }
